refactor: Inline TabModel into CharacterModel
