+++ common/src/main/java/com/wynntils/models/character/type/StatusEffect.java
@@ -2,20 +2,26 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.character.statuseffects;
+package com.wynntils.models.character.type;
 
-public class StaticStatusTimer extends StatusTimer {
+public class StatusEffect {
+    protected final String fullName;
+    private final String name; // The name of the consumable (also used to identify it)
+    protected String displayedTime; // The displayed time remaining. Allows for xx:xx for infinite time effects.
+    protected String prefix; // The prefix to display before the name. Not included in identifying name.
 
-    private String displayedTime; // The displayed time remaining. Allows for xx:xx for infinite time effects.
-    private String prefix; // The prefix to display before the name. Not included in identifying name.
-
-    private final String fullName;
-
-    public StaticStatusTimer(String prefix, String name, String displayedTime) {
-        super(name);
-        this.prefix = prefix;
+    public StatusEffect(String name, String displayedTime, String prefix) {
+        this.name = name;
         this.displayedTime = displayedTime;
+        this.prefix = prefix;
         this.fullName = getPrefix() + " " + getName() + " " + getDisplayedTime();
+    }
+
+    /**
+     * @return The name of the consumable
+     */
+    public String getName() {
+        return name;
     }
 
     /**
@@ -46,7 +52,6 @@ public class StaticStatusTimer extends StatusTimer {
         this.prefix = prefix;
     }
 
-    @Override
     public String asString() {
         return fullName;
     }
